GOSemSimDATA-class {GOSemSim}	R Documentation

Class "GOSemSimDATA" This class stores IC and gene to go mapping for semantic similarity measurement

Description

Class "GOSemSimDATA" This class stores IC and gene to go mapping for semantic similarity measurement

Slots

keys

gene ID

ont

ontology

IC

IC data

geneAnno

gene to GO mapping

tcssdata

tcssdata

metadata

metadata

[Package GOSemSim version 2.20.0 Index]